+++ src/parser/enrichers/item/NeedlerPistol.mjs
@@ -22,9 +22,9 @@ export default class NeedlerPistol extends DDBEnricherData {
   get documentStub() {
     return {
       documentType: "consumable",
-      parsingType: "wonderous",
+      parsingType: "wondrous",
       systemType: {
-        value: "trinket",
+        value: "wondrous",
       },
     };
   }